site-server/requirements.txt
# PROD
fastapi==0.115.5
uvicorn==0.32.0
SQLAlchemy==2.0.36
python-dotenv==1.0.1
asyncpg==0.30.0
alembic==1.14.0
pydantic==2.9.2
pydantic-settings==2.6.1
bcrypt==4.2.1
sqladmin==0.20.1
itsdangerous==2.2.0
Pillow==11.1.0

# DEV
docker==7.1.0
pytest==8.3.3
pytest-alembic==0.11.1
pytest-asyncio==0.24.0
Faker==33.1.0
